practice-projects/flask-mvc/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.29
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
